Screener - ICT Market Structure Shift
Disclaimer:
The information and publications are not intended to be, and do not constitute, financial, investment, trading, or any other form of advice or recommendations supplied or endorsed by the publisher or the creator of this script. Users are advised to exercise their own judgment and seek professional advice before making any financial decisions.

Open-source script:
This repository contains a script created by the original author, shared openly under the terms of the Mozilla Public License 2.0. The license details can be found at https://mozilla.org/MPL/2.0/.
By making the script open-source, the author encourages collaboration and scrutiny, allowing users to understand and verify the underlying indicators.

Indicator Description:
Title: Screener - ICT Market Structure Shift

Introduction:
This screener is designed based on the "Market Structure Shift" concept introduced by Michael Joe Huddleston, popularly known as "ICT," and extensively covered in the "2016 Premium ICT Mentorship Core Content" lectures available on his YouTube channel, "The Inner Circle Trader."
Market Structure Shift: For a bullish scenario, the indicator identifies instances where the price breaks past a swing low and subsequently moves upward, closing above the previous swing high. This occurrence may signal a potential trend reversal. Conversely, for a bearish view, the opposite conditions apply.
Due to the interrelated nature of various topics and jargon, it is recommended to watch the associated videos for a more comprehensive grasp of entry and exit rules.

What is this Indicator?:
The indicator serves as a tool to filter stocks based on the Market Structure Shift concept, aiming to highlight potential trading opportunities in line with the principles taught by ICT.
It is an extension of the "Quasimodo Pattern by EmreKb" indicator, building on its functionality to enhance the screening process.

Screener:
Screens up to 40 symbols, adhering to TradingView's limitation of 40 requests per indicator.
The total character count for the indicator cannot exceed 4096 characters.
Utilize a comma as a delimiter without any spaces when inputting symbols (e.g., 'EXCHANGE:SYMBOL_A,EXCHANGE:SYMBOL_B').

How to add the Code to TradingView Chart?:
1. Open TradingView chart.
2. Click "Pine Editor" at the bottom.
3. Copy indicator code and Paste it in Pine Editor.
4. Click "Save" in Pine Editor.
5. Click "Add to Chart" in Pine Editor.
6. Your indicator will be displayed on the chart.